backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy==2.0.36
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
reportlab==4.0.4
schedule==1.2.0
apscheduler==3.10.4
aiohttp==3.9.1